% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_frames.R
\name{extract_frames}
\alias{extract_frames}
\title{Extract frames}
\usage{
extract_frames(path, fps = 1)
}
\arguments{
\item{path}{Path to the folder containing video clips}

\item{fps}{Default: 1}
}
\value{
Directory "images" with the frames from the specified folder.
}
\description{
This function takes a path to a folder containing video clips and returns a folder with the frames.
}
