% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_videos.R
\name{trim_video}
\alias{trim_video}
\title{Trim video}
\usage{
trim_video(video_path, dftrim, output_dir)
}
\arguments{
\item{video_path}{Path to the video}

\item{dftrim}{Data frame containing the periods for each clip}

\item{output_dir}{Directory where the clips will be saved}
}
\value{
Directory with the clips from the specified video.
}
\description{
This function takes a .mp4 video and a data frame with 2 columns named \code{begin} and \code{end} containing the start and end times
for each clip in the format HH:MM:SS and returns a folder with the trimmed videos.
}
\details{
Note: To use this function, you need to have \code{ffmpeg} installed. To install it, run the command
\code{sudo apt-get install ffmpeg} on your terminal.
}
